Moved logic to execute query, from CPT to Executer
