+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/CustomPostTypes/GraphQLPersistedQueryEndpointCustomPostType.php
@@ -19,7 +19,6 @@ use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\AbstractGraphQLEndpointCustom
 use GraphQLAPI\GraphQLAPI\Services\Helpers\BlockContentHelpers;
 use GraphQLAPI\GraphQLAPI\Services\Helpers\GraphQLQueryPostTypeHelpers;
 use GraphQLAPI\GraphQLAPI\Services\Taxonomies\GraphQLQueryTaxonomy;
-use GraphQLByPoP\GraphQLRequest\Hooks\VarsHookSet as GraphQLRequestVarsHooks;
 use PoP\ComponentModel\Instances\InstanceManagerInterface;
 use PoP\Hooks\HooksAPIInterface;
 use WP_Post;
@@ -243,19 +242,6 @@ class GraphQLPersistedQueryEndpointCustomPostType extends AbstractGraphQLEndpoin
         return $content;
     }
 
-    /**
-     * Provide the query to execute and its variables
-     *
-     * @return mixed[] Array with 2 elements: [$graphQLQuery, $graphQLVariables]
-     */
-    protected function getGraphQLQueryAndVariables(?WP_Post $graphQLQueryPost): array
-    {
-        /**
-         * Extract the query from the post (or from its parents), and set it in $vars
-         */
-        return $this->graphQLQueryPostTypeHelpers->getGraphQLQueryPostAttributes($graphQLQueryPost, true);
-    }
-
     protected function getEndpointOptionsBlock(): AbstractEndpointOptionsBlock
     {
         /**
@@ -281,42 +267,5 @@ class GraphQLPersistedQueryEndpointCustomPostType extends AbstractGraphQLEndpoin
 
         // `true` is the default option in Gutenberg, so it's not saved to the DB!
         return $optionsBlockDataItem['attrs'][PersistedQueryEndpointOptionsBlock::ATTRIBUTE_NAME_ACCEPT_VARIABLES_AS_URL_PARAMS] ?? $default;
-    }
-
-    /**
-     * Check if requesting the single post of this CPT and, in this case, set the request with the needed API params
-     *
-     * @param array<array> $vars_in_array
-     */
-    public function addGraphQLVars(array $vars_in_array): void
-    {
-        if (\is_singular($this->getCustomPostType())) {
-            // Check if it is enabled, by configuration
-            [&$vars] = $vars_in_array;
-            if (!$this->isEndpointEnabled($vars['routing-state']['queried-object-id'])) {
-                return;
-            }
-
-            /** @var GraphQLRequestVarsHooks */
-            $graphQLAPIRequestHookSet = $this->instanceManager->getInstance(GraphQLRequestVarsHooks::class);
-
-            // The Persisted Query is also standard GraphQL
-            $graphQLAPIRequestHookSet->setStandardGraphQLVars($vars);
-
-            // Remove the VarsHookSet from the GraphQLRequest, so it doesn't process the GraphQL query
-            // Otherwise it will add error "The query in the body is empty"
-            /**
-             * @var callable
-             */
-            $action = [$graphQLAPIRequestHookSet, 'addVars'];
-            \remove_action(
-                'ApplicationState:addVars',
-                $action,
-                20
-            );
-
-            // Execute the original logic
-            parent::addGraphQLVars($vars_in_array);
-        }
     }
 }
