+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/EndpointExecuters/CustomEndpointGraphQLQueryResolutionEndpointExecuter.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Services\EndpointExecuters;
 
+use WP_Post;
 use GraphQLAPI\GraphQLAPI\Registries\ModuleRegistryInterface;
 use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\AbstractGraphQLEndpointCustomPostType;
 use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\GraphQLCustomEndpointCustomPostType;
+use GraphQLByPoP\GraphQLRequest\Execution\QueryExecutionHelpers;
 use PoP\ComponentModel\Instances\InstanceManagerInterface;
 
 class CustomEndpointGraphQLQueryResolutionEndpointExecuter extends AbstractGraphQLQueryResolutionEndpointExecuter implements CustomEndpointExecuterServiceTagInterface
@@ -25,5 +27,18 @@ class CustomEndpointGraphQLQueryResolutionEndpointExecuter extends AbstractGraph
     protected function getCustomPostType(): AbstractGraphQLEndpointCustomPostType
     {
         return $this->graphQLCustomEndpointCustomPostType;
+    }
+
+    /**
+     * Provide the query to execute and its variables
+     *
+     * @return mixed[] Array of 2 elements: [query, variables]
+     */
+    protected function getGraphQLQueryAndVariables(?WP_Post $graphQLQueryPost): array
+    {
+        /**
+         * Extract the query from the BODY through standard GraphQL endpoint execution
+         */
+        return QueryExecutionHelpers::extractRequestedGraphQLQueryPayload();
     }
 }
